updated percentage in poll option

+++ internalsdk/src/main/java/com/likeminds/internalsdk/post/model/_PollOption_.kt
@@ -8,7 +8,7 @@ class _PollOption_ private constructor(
     @SerializedName("is_selected")
     val isSelected: Boolean,
     @SerializedName("percentage")
-    val percentage: Int,
+    val percentage: Float,
     @SerializedName("text")
     val uuid: String,
     @SerializedName("vote_count")
@@ -17,13 +17,13 @@ class _PollOption_ private constructor(
     class Builder {
         private var id: String = ""
         private var isSelected: Boolean = false
-        private var percentage: Int = 0
+        private var percentage: Float = 0f
         private var uuid: String = ""
         private var voteCount: Int = 0
 
         fun id(id: String) = apply { this.id = id }
         fun isSelected(isSelected: Boolean) = apply { this.isSelected = isSelected }
-        fun percentage(percentage: Int) = apply { this.percentage = percentage }
+        fun percentage(percentage: Float) = apply { this.percentage = percentage }
         fun uuid(uuid: String) = apply { this.uuid = uuid }
         fun voteCount(voteCount: Int) = apply { this.voteCount = voteCount }
