+++ likemindsfeed/src/main/java/com/likeminds/likemindsfeed/post/model/PollOption.kt
@@ -3,20 +3,20 @@ package com.likeminds.likemindsfeed.post.model
 class PollOption private constructor(
     val id: String,
     val isSelected: Boolean,
-    val percentage: Int,
+    val percentage: Float,
     val uuid: String,
     val voteCount: Int
 ) {
     class Builder {
         private var id: String = ""
         private var isSelected: Boolean = false
-        private var percentage: Int = 0
+        private var percentage: Float = 0f
         private var uuid: String = ""
         private var voteCount: Int = 0
 
         fun id(id: String) = apply { this.id = id }
         fun isSelected(isSelected: Boolean) = apply { this.isSelected = isSelected }
-        fun percentage(percentage: Int) = apply { this.percentage = percentage }
+        fun percentage(percentage: Float) = apply { this.percentage = percentage }
         fun uuid(uuid: String) = apply { this.uuid = uuid }
         fun voteCount(voteCount: Int) = apply { this.voteCount = voteCount }
